drew's initial commit to add Chess War Features to HexChessWar.
This code is not in a working state to beware.

I felt it was best to commit it now before I break something worse
and lose the progress I have made so far.

Plus, it may be helpful to see where this is going in any case
even if things are not working yet.

 Changes to be committed:
	modified:   LICENSE
	modified:   README.md
	new file:   notworking_notes.txt
	modified:   project.godot
	modified:   scenes/board.tscn
	modified:   scenes/menu.tscn
	modified:   scr/Board.gd
	modified:   scr/Game.gd
	modified:   scr/Mapping.gd
	modified:   scr/Menu.gd
	modified:   scr/Piece.gd
	modified:   scr/PlayersData.gd