Well, I have messed up a bunch of stuff Egor had working and likely some stuff I had working as well.
However, I think I have a multiplayer, synced, Level 2 battle
happening. It is not perfect, but I think it might be playable
as a demo.

Please try it and give feedback.

The Alert stuff is  from Andy over at godot-chess-war.

Consider the licenses to be what they need to be to make this code
blending work. I have to figure that out but for now I am trying
to create semi working code...
On branch main
 Changes to be committed:
	new file:   scenes/Alert.tscn
	modified:   scenes/board.tscn
	modified:   scenes/menu.tscn
	new file:   scr/Alert.gd
	modified:   scr/Game.gd
	modified:   scr/Lobby.gd
	modified:   scr/Multiplayer.gd